add unit tests to cover the pod network ports binding
